+++ src/Exception.php
@@ -5,6 +5,6 @@ namespace Issei\Spike;
 /**
  * @author Issei Murasawa <issei.m7@gmail.com>
  */
-class Exception extends \Exception
+class Exception extends \RuntimeException
 {
 }